feat: api 코드 이동 및 변경 #30
- instance 추가
- fetcher -> httpClient 로 변경
- httpClient 파일 utils 폴더로 이동
